requirements.txt
biopython==1.83
ete3==3.1.3
numpy==1.26.4
six==1.16.0
